feat(cmd/gpud): move "join" command to a separate package
