+++ cmd/gpud/join/command.go
@@ -1,4 +1,5 @@
-package command
+// Package join implements the "join" command.
+package join
 
 import (
 	"bufio"
@@ -28,12 +29,13 @@ import (
 	"github.com/leptonai/gpud/pkg/sqlite"
 )
 
-func cmdJoin(cliContext *cli.Context) (retErr error) {
+func Command(cliContext *cli.Context) (retErr error) {
 	if err := osutil.RequireRoot(); err != nil {
 		return err
 	}
 
-	// Set up logging
+	logLevel := cliContext.String("log-level")
+	logFile := cliContext.String("log-file")
 	zapLvl, err := log.ParseLogLevel(logLevel)
 	if err != nil {
 		return err
